+++ src/lib.rs
@@ -71,6 +71,14 @@ pub struct Signature(ffi::Signature);
 #[derive(Copy, Clone, PartialEq, Eq, Debug)]
 pub struct RecoverableSignature(ffi::RecoverableSignature);
 
+/// Trait describing something that promises to be a 32-byte random number; in particular,
+/// it has negligible probability of being zero or overflowing the group order. Such objects
+/// may be converted to `Message`s without any error paths.
+pub trait ThirtyTwoByteHash {
+    /// Converts the object into a 32-byte array
+    fn into_32(self) -> [u8; 32];
+}
+
 impl RecoveryId {
     #[inline]
     /// Allows library users to create valid recovery IDs from i32.
@@ -350,6 +358,12 @@ impl Message {
             }
             _ => Err(Error::InvalidMessage)
         }
+    }
+
+    #[inline]
+    /// Converts a 32-byte hash directly to a message without error paths
+    pub fn from_hash<T: ThirtyTwoByteHash>(t: T) -> Message {
+        Message(t.into_32())
     }
 }